Extract `inferParserByLanguage`
